fixed the sidebar
